MicroService/auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
python-jose
minio
psycopg2-binary
python-dotenv
python-multipart
bcrypt